Moved path tools src files to subfolder
