+++ addons/PathMesh3D/src/path_tools/path_static_body_3d.hpp
