+++ addons/PathMesh3D/src/register_types.cpp
@@ -1,14 +1,14 @@
-#include "path_mesh_3d.hpp"
-#include "path_extrude_3d.hpp"
-#include "path_multimesh_3d.hpp"
-#include "path_scene_3d.hpp"
+#include "path_tools/path_mesh_3d.hpp"
+#include "path_tools/path_extrude_3d.hpp"
+#include "path_tools/path_multimesh_3d.hpp"
+#include "path_tools/path_scene_3d.hpp"
 #include "path_modifier_3d.hpp"
 #include "path_extrude_profile_base.hpp"
 #include "extrude_profiles/path_extrude_profile_manual.hpp"
 #include "extrude_profiles/path_extrude_profile_rect.hpp"
 #include "extrude_profiles/path_extrude_profile_circle.hpp"
-#include "path_static_body_3d.hpp"
-#include "path_area_3d.hpp"
+#include "path_tools/path_static_body_3d.hpp"
+#include "path_tools/path_area_3d.hpp"
 #include "register_types.hpp"
 
 void initialize_path_mesh_3d(ModuleInitializationLevel p_level) {
